+++ src/main/java/com/kruemblegard/client/ModEntityRenderers.java
@@ -31,5 +31,8 @@ public class ModEntityRenderers {
 
         event.registerEntityRenderer(ModProjectileEntities.ARCANE_STORM.get(),
                 ctx -> new GenericProjectileRenderer<>(ctx, 0.5f));
+
+        event.registerEntityRenderer(ModProjectileEntities.TRAPROCK_STONE.get(),
+                ctx -> new GenericProjectileRenderer<>(ctx, 0.5f));
     }
 }